'-Wall', ASan leak fixes, connection close on unspecified bodies
This adds '-Wall', '-Wextra', '-Werror' compiler flags and fixes the
associated warnings.

It also fixes a few leaks detected by ASan.

It also adds unit tests to chunked-encoded parsing and makes some
adjustements accordingly. I don't believe, in the end, this changed much
behaviorally, but it could in theory make the client read the entire
final chunk (the zero-body one) when, before, it could have stopped
reading that after the entire expected body was received. But any
previous "bad" behavior would require specific read behavior such that
the final chunk was on a different read than the rest, which maybe is
unlikely. And even then, I'm not sure if something bad would have
happened.

The big behavioral change is having the server close the connection
after the entire writing of the body is done and there is no
content-length. This prevents timeouts from the client (generally, the
proxy actually) as it waits for the unspecified body.