Write HTTP fields in the yaml-specified order.
Before this change, fields were written to the socket in the order
dictated when iterating over them via an unordred_multimap. This meant
there was no guaranteed order, and the order was different in the
unordered_multimap RHEL 7 g++ and mac clang implementations. This change
makes it so that the fields are written in the order specified by the
yaml file.

Sending the fields in the order specified by the YAML file is a useful
feature for the user and greatly simplifies the AuTests which no longer
need to branch their expectations based upon system information.