Update server to be able to derive key from just client-request
Before this, the server had to get the key from a proxy-request node. If
an all:headers::fields::uid existed, it happened to get it from there,
but otherwise the server could not derive a key for a transaction
without a proxy-request. We don't want the user to be required to add
a proxy-request node, so this wasn't acceptable.