Fixing memory leak of TLSSession::server_context
Fixing a memory leak found with ASan. The call to H2Session::init called
TLSSession::init with server_context after TLSSession::init already
called it for the same server_context. (Both TLSSession and H2Session
share the same server_context because they need not be different between
the two.) This caused server_context to be initialized twice with
SSL_CTX_new, with the first one being immediately overwritten by the
second, thus causing the first to be leaked.