Updating the HTTP library dependencies.
This updates the OpenSSL and HTTP library dependencies as they have
changed. The nghttp2 commit that was previously referenced got removed
due to a rebase in that repository, so that needed to be updated. Also,
curl and ATS have moved to using quictls/openssl instead of
tatsuhiro-t/openssl for OpenSSL. This updates that as well.