HTTP/2 timing fix for --rate (#176)
We had a sleep_for to space out HTTP/2 transactions (streams) within a
session to adhere to the --rate parameter. This caused delays, though,
for reading responses for other streams. The sleep_for is unnecessary:
the reading from the socket has a timeout that corresponds to our delay
anyway. Simply remove the sleep_for to fix the timing issue.