Test Proxy: switching to httpx from hyper.
hyper, heretofore our test proxy's HTTP/2 package, is no longer
supported and does not work with Python 3.10. We therefore need to
switch it out for a different HTTP/2 Python package. httpx is the
recommended replacement by the final commit in hyper and seems to work
OK.

Switch to httpx goes pretty well, but the server gold file expectations
need to be updated because httpx is "smarter" about accept-*,
user-agent, host headers, and the like.