Update http2 test proxy to use the h2 lib (#265)
Currently the http2 test proxy uses the httpx library, which works great in general. However it doesn't support trailer header. As we are working on adding trailer header support in Proxy Verifier, it would be good to have the test proxy support trailer header, so that we can verify the proper functionality of the feature.

This PR updates the http2 test proxy, replacing the usage of httpx with the python-hyper/h2 library. The h2 is a lower-level library that supports trailer header and is used in the client-facing logic already. Verified that after these changes(with adjusted expectations), all existing tests pass.

Note that while this PR updates the test proxy to use a library that can support trailer header, the actual trailer header handling logic in the proxy would be added in another PR(on top of this one) for clarity and convenience of testing. The changes in this PR makes it quite trivial to add that.